refactor(card-editor): regroup composables and adopt client_id keying
Move card-list-controller, card-mutations, card-selection, and
virtual-card-list under src/composables/card-editor/ and rework the
internals:

- Each composable now owns the queries/mutations it needs from
  @/api; controller passes only deck_id, decoupling card-mutations
  from VirtualCardList.
- virtual-card-list keys every rendered card on a stable client_id
  (uid()-based) instead of branching on the sign of card.id; promote
  seeds the persisted-id Map so the v-for key survives the
  temp -> persisted handoff and the text editor stays mounted.
- controller intent handlers split into single-responsibility helpers
  (withSaving, insertTemp, confirmDelete, afterDelete,
  resolveDeleteArgs, openMoveModal); orchestrators stay flat with
  early returns and at most one level of nesting.
- card-mutations becomes a thin wrapper layer over @/api/cards
  hooks (insertCard, saveCard, deleteCards, moveCards) with no
  knowledge of list state.
- card-selection reads total_card_count from useDeckQuery's
  card_count, dropping the dead useDeckCardIdsQuery + the
  fetchCardIdsByDeckId db helper.
- Consumers (list, list-item, list-item-card, card-grid, grid-item,
  bulk-select-toolbar, card-importer, deck-view, card components)
  re-pointed at the new module paths and updated to key v-for on
  card.client_id.

Tests follow the source move (tests/unit/composables/card-editor/),
new mocks for the thin mutation wrappers, additional coverage for
the collectCards "id already in selection" branch and the
selectCard "id not in deselected" no-op branch.
