+++ src/api/cards/queries/index.ts
@@ -1,5 +1,4 @@
 export * from './cards-in-deck'
 export * from './cards-page'
-export * from './card-ids'
 export * from './search-in-deck'
 export * from './member-card-count'
