+++ src/views/deck/card-editor/index.vue
@@ -3,7 +3,7 @@ import List from './list.vue'
 import { useI18n } from 'vue-i18n'
 import UiButton from '@/components/ui-kit/button.vue'
 import { inject } from 'vue'
-import { type CardListController } from '@/composables/card-list-controller'
+import { type CardListController } from '@/composables/card-editor/card-list-controller'
 
 const { t } = useI18n()
 const { all_cards, addCard } = inject<CardListController>('card-editor')!
