+++ src/views/deck/card-editor/list.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
 import ListItem from './list-item.vue'
 import { inject, useTemplateRef } from 'vue'
-import { type CardListController } from '@/composables/card-list-controller'
+import { type CardListController } from '@/composables/card-editor/card-list-controller'
 
-const { all_cards, getKey, hasNextPage, isLoading, observeSentinel } =
+const { all_cards, hasNextPage, isLoading, observeSentinel } =
   inject<CardListController>('card-editor')!
 
 const sentinel = useTemplateRef<HTMLElement>('sentinel')
@@ -15,7 +15,7 @@ observeSentinel(sentinel)
   <div data-testid="card-list" class="w-full pt-6 flex flex-col items-center">
     <list-item
       v-for="(card, index) in all_cards"
-      :key="getKey(card)"
+      :key="card.client_id"
       :index="index"
       :card="card"
       :duplicate="card.is_duplicate ?? false"
