+++ tests/unit/api/cards/cards-page.test.js
@@ -15,7 +15,7 @@ vi.mock('@/supabase-client', () => ({
 
 vi.mock('@/utils/logger', () => ({ default: { error: vi.fn() } }))
 
-import { fetchCardsPageByDeckId, fetchCardIdsByDeckId } from '@/api/cards/db/cards-page'
+import { fetchCardsPageByDeckId } from '@/api/cards/db/cards-page'
 
 beforeEach(() => {
   fromMock.mockClear()
@@ -52,28 +52,5 @@ describe('fetchCardsPageByDeckId', () => {
     await expect(fetchCardsPageByDeckId({ deck_id: 10, offset: 0, limit: 50 })).rejects.toThrow(
       'boom'
     )
-  })
-})
-
-describe('fetchCardIdsByDeckId', () => {
-  test('selects only the id column from cards, ordered by rank', async () => {
-    // fetchCardIdsByDeckId doesn't call .range — it terminates on .order, which
-    // must resolve as a thenable. Override orderMock for this describe block.
-    orderMock.mockImplementationOnce(() =>
-      Promise.resolve({ data: [{ id: 1 }, { id: 2 }], error: null })
-    )
-    const ids = await fetchCardIdsByDeckId(10)
-    expect(fromMock).toHaveBeenCalledWith('cards')
-    expect(selectMock).toHaveBeenCalledWith('id')
-    expect(eqMock).toHaveBeenCalledWith('deck_id', 10)
-    expect(orderMock).toHaveBeenCalledWith('rank', { ascending: true })
-    expect(ids).toEqual([1, 2])
-  })
-
-  test('throws when the request errors', async () => {
-    orderMock.mockImplementationOnce(() =>
-      Promise.resolve({ data: null, error: { message: 'denied' } })
-    )
-    await expect(fetchCardIdsByDeckId(10)).rejects.toThrow('denied')
   })
 })
