+++ tests/unit/composables/card-editor/card-selection.test.js
@@ -1,10 +1,20 @@
-import { describe, test, expect } from 'vite-plus/test'
+import { describe, test, expect, vi } from 'vite-plus/test'
 import { ref } from 'vue'
 import { card } from '@tests/fixtures/card'
-import { useCardSelection } from '@/composables/card-selection'
 
-function makeIdsQuery(ids = []) {
-  return { data: ref(ids) }
+const { useDeckQueryMock } = vi.hoisted(() => ({
+  useDeckQueryMock: vi.fn()
+}))
+
+vi.mock('@/api/decks', () => ({
+  useDeckQuery: useDeckQueryMock
+}))
+
+import { useCardSelection } from '@/composables/card-editor/card-selection'
+
+function makeSelection(ids = []) {
+  useDeckQueryMock.mockReturnValueOnce({ data: ref({ id: 10, card_count: ids.length }) })
+  return useCardSelection(10)
 }
 
 function makeCard(overrides = {}) {
@@ -16,20 +26,21 @@ describe('useCardSelection', () => {
 
   describe('initial state', () => {
     test('starts with empty selection and not in select-all mode', () => {
-      const sel = useCardSelection(makeIdsQuery([]))
+      const sel = makeSelection([])
       expect(sel.selected_card_ids.value).toEqual([])
       expect(sel.deselected_ids.value).toEqual([])
       expect(sel.select_all_mode.value).toBe(false)
       expect(sel.selected_count.value).toBe(0)
     })
 
-    test('total_card_count tracks the ids_query length', () => {
-      const sel = useCardSelection(makeIdsQuery([1, 2, 3, 4]))
+    test('total_card_count tracks deck.card_count', () => {
+      const sel = makeSelection([1, 2, 3, 4])
       expect(sel.total_card_count.value).toBe(4)
     })
 
-    test('total_card_count is 0 when the query has no data', () => {
-      const sel = useCardSelection({ data: ref(undefined) })
+    test('total_card_count is 0 when the deck query has no data', () => {
+      useDeckQueryMock.mockReturnValueOnce({ data: ref(undefined) })
+      const sel = useCardSelection(10)
       expect(sel.total_card_count.value).toBe(0)
     })
   })
@@ -38,14 +49,14 @@ describe('useCardSelection', () => {
 
   describe('positive selection', () => {
     test('selectCard adds an id once (no duplicates)', () => {
-      const sel = useCardSelection(makeIdsQuery([1]))
+      const sel = makeSelection([1])
       sel.selectCard(1)
       sel.selectCard(1)
       expect(sel.selected_card_ids.value).toEqual([1])
     })
 
     test('deselectCard removes an id', () => {
-      const sel = useCardSelection(makeIdsQuery([1, 2]))
+      const sel = makeSelection([1, 2])
       sel.selectCard(1)
       sel.selectCard(2)
       sel.deselectCard(1)
@@ -53,13 +64,13 @@ describe('useCardSelection', () => {
     })
 
     test('deselectCard is a no-op for an id not in the selection', () => {
-      const sel = useCardSelection(makeIdsQuery([1]))
+      const sel = makeSelection([1])
       sel.deselectCard(99)
       expect(sel.selected_card_ids.value).toEqual([])
     })
 
     test('toggleSelectCard flips selection state', () => {
-      const sel = useCardSelection(makeIdsQuery([1]))
+      const sel = makeSelection([1])
       sel.toggleSelectCard(1)
       expect(sel.isCardSelected(1)).toBe(true)
       sel.toggleSelectCard(1)
@@ -67,21 +78,21 @@ describe('useCardSelection', () => {
     })
 
     test('isCardSelected mirrors the positive list', () => {
-      const sel = useCardSelection(makeIdsQuery([1, 2]))
+      const sel = makeSelection([1, 2])
       sel.selectCard(1)
       expect(sel.isCardSelected(1)).toBe(true)
       expect(sel.isCardSelected(2)).toBe(false)
     })
 
     test('selected_count tracks the positive list length', () => {
-      const sel = useCardSelection(makeIdsQuery([1, 2]))
+      const sel = makeSelection([1, 2])
       sel.selectCard(1)
       sel.selectCard(2)
       expect(sel.selected_count.value).toBe(2)
     })
 
     test('all_cards_selected is true when the positive list matches the deck total', () => {
-      const sel = useCardSelection(makeIdsQuery([1, 2]))
+      const sel = makeSelection([1, 2])
       expect(sel.all_cards_selected.value).toBe(false)
       sel.selectCard(1)
       sel.selectCard(2)
@@ -89,7 +100,7 @@ describe('useCardSelection', () => {
     })
 
     test('all_cards_selected is false when the deck is empty (no vacuous true)', () => {
-      const sel = useCardSelection(makeIdsQuery([]))
+      const sel = makeSelection([])
       expect(sel.all_cards_selected.value).toBe(false)
     })
   })
@@ -98,7 +109,7 @@ describe('useCardSelection', () => {
 
   describe('select-all mode', () => {
     test('selectAllCards flips into select-all mode and clears the positive list', () => {
-      const sel = useCardSelection(makeIdsQuery([1]))
+      const sel = makeSelection([1])
       sel.selectCard(1)
       sel.selectAllCards()
       expect(sel.select_all_mode.value).toBe(true)
@@ -106,7 +117,7 @@ describe('useCardSelection', () => {
     })
 
     test('isCardSelected is true for every id except those deselected', () => {
-      const sel = useCardSelection(makeIdsQuery([1, 2]))
+      const sel = makeSelection([1, 2])
       sel.selectAllCards()
       sel.deselectCard(1)
       expect(sel.isCardSelected(1)).toBe(false)
@@ -114,7 +125,7 @@ describe('useCardSelection', () => {
     })
 
     test('deselectCard accumulates ids once in the deselected list', () => {
-      const sel = useCardSelection(makeIdsQuery([1, 2]))
+      const sel = makeSelection([1, 2])
       sel.selectAllCards()
       sel.deselectCard(1)
       sel.deselectCard(1)
@@ -122,15 +133,24 @@ describe('useCardSelection', () => {
     })
 
     test('selectCard in select-all mode removes the id from the deselected list', () => {
-      const sel = useCardSelection(makeIdsQuery([1]))
+      const sel = makeSelection([1])
       sel.selectAllCards()
       sel.deselectCard(1)
       sel.selectCard(1)
       expect(sel.deselected_ids.value).toEqual([])
     })
 
+    test('selectCard in select-all mode is a no-op when the id is not in the deselected list', () => {
+      const sel = makeSelection([1, 2])
+      sel.selectAllCards()
+      sel.deselectCard(1)
+      sel.selectCard(2)
+      expect(sel.deselected_ids.value).toEqual([1])
+      expect(sel.selected_card_ids.value).toEqual([])
+    })
+
     test('selected_count = total - deselected in select-all mode', () => {
-      const sel = useCardSelection(makeIdsQuery([1, 2, 3]))
+      const sel = makeSelection([1, 2, 3])
       sel.selectAllCards()
       expect(sel.selected_count.value).toBe(3)
       sel.deselectCard(2)
@@ -138,7 +158,7 @@ describe('useCardSelection', () => {
     })
 
     test('selected_count is clamped to zero if deselected exceeds total (defensive)', () => {
-      const sel = useCardSelection(makeIdsQuery([1]))
+      const sel = makeSelection([1])
       sel.selectAllCards()
       sel.deselectCard(1)
       sel.deselectCard(99)
@@ -146,7 +166,7 @@ describe('useCardSelection', () => {
     })
 
     test('all_cards_selected is true iff nothing has been deselected', () => {
-      const sel = useCardSelection(makeIdsQuery([1, 2]))
+      const sel = makeSelection([1, 2])
       sel.selectAllCards()
       expect(sel.all_cards_selected.value).toBe(true)
       sel.deselectCard(1)
@@ -154,7 +174,7 @@ describe('useCardSelection', () => {
     })
 
     test('clearSelectedCards exits select-all mode and empties both lists', () => {
-      const sel = useCardSelection(makeIdsQuery([1]))
+      const sel = makeSelection([1])
       sel.selectAllCards()
       sel.deselectCard(1)
       sel.clearSelectedCards()
@@ -164,13 +184,13 @@ describe('useCardSelection', () => {
     })
 
     test('toggleSelectAll enters select-all when nothing is selected', () => {
-      const sel = useCardSelection(makeIdsQuery([1, 2]))
+      const sel = makeSelection([1, 2])
       sel.toggleSelectAll()
       expect(sel.select_all_mode.value).toBe(true)
     })
 
     test('toggleSelectAll exits when everything is already selected', () => {
-      const sel = useCardSelection(makeIdsQuery([1]))
+      const sel = makeSelection([1])
       sel.toggleSelectAll() // enter select-all
       sel.toggleSelectAll() // back out
       expect(sel.select_all_mode.value).toBe(false)
@@ -183,7 +203,7 @@ describe('useCardSelection', () => {
   describe('filterSelected', () => {
     test('returns only cards whose id is currently selected (positive mode)', () => {
       const cards = [makeCard({ id: 1 }), makeCard({ id: 2 }), makeCard({ id: 3 })]
-      const sel = useCardSelection(makeIdsQuery([1, 2, 3]))
+      const sel = makeSelection([1, 2, 3])
       sel.selectCard(1)
       sel.selectCard(3)
       expect(sel.filterSelected(cards).map((c) => c.id)).toEqual([1, 3])
@@ -191,7 +211,7 @@ describe('useCardSelection', () => {
 
     test('returns all cards except deselected ids in select-all mode', () => {
       const cards = [makeCard({ id: 1 }), makeCard({ id: 2 })]
-      const sel = useCardSelection(makeIdsQuery([1, 2]))
+      const sel = makeSelection([1, 2])
       sel.selectAllCards()
       sel.deselectCard(2)
       expect(sel.filterSelected(cards).map((c) => c.id)).toEqual([1])
@@ -199,7 +219,7 @@ describe('useCardSelection', () => {
 
     test('skips cards with undefined id (defensive)', () => {
       const cards = [{ id: undefined }, makeCard({ id: 1 })]
-      const sel = useCardSelection(makeIdsQuery([1]))
+      const sel = makeSelection([1])
       sel.selectCard(1)
       expect(sel.filterSelected(cards).map((c) => c.id)).toEqual([1])
     })
